Adds the check_files() function to the script
The check_files function verifies that all files in the working directory are of the same file type. It does this by creating a list of all file extensions in the directory. Then it iterates through the list and compares the extensions of each. If the extensions don't match than the function returns an error message. If they do match, then the rename_files function is called.