Creates a file_list list to later create the history csv
Changed the "ext_list" to a file_list where the entire file extension is added to the list. The file check then has to use regex to compare just the extension from each file to validate if it's the same. Now the entire file_list can be passed to a new function which will create the name history.