Creates a file name demo.
Adds a feature which previews the first 5 file name changes. First it iterates through to get the string length of each file name and sets a variable for the longest length. This is used to format the demo names.

Then it iterates through the first 5 names again and previews the old names converted to the new names. Lastly, it asks if the user would like to continue with the file change. Currently, the file change does not occur yet.