Creates a file_history.csv file with old and new names.
A lot of changes were made. Old and new name lists are created before checking the file extensions. Both lists are now used to generate the name change demo. Then the lists are pushed to the new create_history function. They're combined into a dictionary and then printed to the csv file in rows.