Completes file_history creation and File renaming
File renaming now uses the new_name list to rename all of the files. Two "if" statements were added to omit the file_history.csv when performing checks. This allows file extensions to be checked even if the .csv already exists.

Also, a check was added to see if the destination directory exists. If it doesn't than it will alert the user.